// src/sdk/portalPoc/Accounts/User.ts
import { LoginHistory } from "./LoginHistory";
import { UserStatus } from "./UserStatus";

export interface User {

    readonly id: string;
    /**
     * The UUID of the account.
     */
    readonly accountId?: string;

    /**
     * Address.
     */
    readonly address?: string;

    /**
     * Creation UTC time RFC3339.
     */
    readonly createdAt?: Date;

    /**
     * A timestamp of the user creation in the storage, in milliseconds.
     */
    readonly creationTime?: number;

    /**
     * The email address.
     */
    readonly email?: string;

    /**
     * A flag indicating whether the user&#39;s email address has been verified or not.
     */
    readonly emailVerified?: boolean;

    /**
     * The full name of the user.
     */
    readonly fullName?: string;

    readonly groups?: Array<string>;

    /**
     * A timestamp of the latest login of the user, in milliseconds.
     */
    readonly lastLoginTime?: number;

    /**
     * Timestamps, succeedings, IP addresses and user agent information of the last five logins of the user, with timestamps in RFC3339 format.
     */
    readonly loginHistory?: Array<LoginHistory>;

    /**
     * A flag indicating that receiving marketing information has been accepted.
     */
    readonly marketingAccepted?: boolean;

    /**
     * The password when creating a new user. It will be generated when not present in the request.
     */
    readonly password?: string;

    /**
     * A timestamp of the latest change of the user password, in milliseconds.
     */
    readonly passwordChangedTime?: number;

    /**
     * Phone number.
     */
    readonly phoneNumber?: string;

    /**
     * The status of the user. ENROLLING state indicates that the user is in the middle of the enrollment process. INVITED means that the user has not accepted the invitation request. RESET means that the password must be changed immediately. INACTIVE users are locked out and not permitted to use the system.
     */
    readonly status?: UserStatus;

    /**
     * A flag indicating that the General Terms and Conditions has been accepted.
     */
    readonly termsAccepted?: boolean;

    /**
     * A flag indicating whether 2-factor authentication (TOTP) has been enabled.
     */
    readonly twoFactorAuthentication?: boolean;

    /**
     * Last update UTC time RFC3339.
     */
    readonly updatedAt?: Date;

    /**
     * A username containing alphanumerical letters and -,._@+= characters.
     */
    readonly username?: string;
}

export interface UserUpdateRequest {
    readonly fullName: string;
    readonly email: string;
}